v0.8 Sprint 3 (QA-018a): decompose CommunityHttp2SessionProcessor — extract 4 pkg-private seams
Decomposes the 896-LOC HTTP/2 session processor by extracting four
self-contained concerns to pkg-private siblings:

- Http2SessionContext (275 LOC): per-connection HPACK encoder/decoder
  pair, frame codec with peer-advertised SETTINGS_MAX_FRAME_SIZE,
  header-block assembler tracking HEADERS/CONTINUATION state, per-stream
  request registry, last-stream-id pointer for GOAWAY framing,
  applyPeerSettings dispatcher, response-headers encoder, and the
  isConnectionSpecificHeader helper (RFC 7540 §8.1.2.2).
- PendingRequestHeaders (94 LOC): HPACK-decode accumulator for one
  HTTP/2 request — enforces RFC 7540 §8.1.2.1 pseudo-header ordering,
  §8.1.2.2 connection-specific header rejection, and the :method/:path
  pseudo-header requirements.
- CommunityHttp2ControlFrames (120 LOC): writers for SETTINGS, SETTINGS-ACK,
  PING-ACK, RST_STREAM (REFUSED/CANCEL), GOAWAY — uniform borrow-encode-write
  shape, parameterized over MemoryAllocator + TransportStream.
- CommunityHttp2FrameFragments (96 LOC): pad/priority field extraction for
  HEADERS and DATA frame payloads (RFC 7540 §6.1 / §6.2).

Residual processor (472 LOC) retains the frame-loop, frame dispatcher
switch, HEADERS/CONTINUATION/DATA handlers, response writer (with the
PERF-061 per-response reusable carrier), and the ingress read loop.

Numbers:
- Processor LOC: 896 → 472 (-47%).
- Class-level @SuppressWarnings: 7 → 5.
  - Closed: PMD.ExcessiveImports, PMD.CouplingBetweenObjects (extracted
    HPACK/codec/assembler types out of outer's import surface).
  - Retained: GodClass (residual WMC=72), TooManyMethods, CyclomaticComplexity
    (frame-dispatcher switch + frame loop), CloseResource (LoanedBuffer
    ownership transfers via response.body() pipeline),
    AvoidCatchingGenericException (frame-loop catches RuntimeException to
    surface PROTOCOL_ERROR via GOAWAY).

Public surface unchanged: CommunityHttp2SessionProcessor pkg-private
constructor signature, handlePriorKnowledge / handleUpgrade entry points,
and all related types (Http2DecodedRequest, Http2RequestStreamState, etc.)
preserved. Callers need no changes.

Verified:
- Full reactor mvn verify -Pcoverage -DskipITs SUCCESS (1:44)
- HTTP test suite 75/76 green (1 pre-existing skip, no new failures)
- PMD + Checkstyle 0 violations
- JaCoCo BUNDLE LINE coverage check met

Sprint 3 GodClass tracker after this PR:
- QA-015 CommunityHttpClientEngine ✅ #133
- QA-016 NativeTcpStream ✅ #134
- QA-017 JdbcFlowSnapshotStore ✅ #136
- QA-018a CommunityHttp2SessionProcessor (this PR) ✅
- QA-018b SubsystemOrchestrator — next (Core/bootstrap, boundary-sensitive)
