+++ exeris-kernel-community/src/main/java/eu/exeris/kernel/community/http/PendingRequestHeaders.java
@@ -1,0 +1,89 @@
+/*
+ * Copyright (C) 2025-2026 Exeris Systems.
+ *
+ * Licensed under the Apache License, Version 2.0 with Commons Clause.
+ * You may use, modify, and distribute this file under those terms.
+ * Commercial resale of this software as a competing product is prohibited.
+ * See LICENSE-COMMUNITY in the repository root for the full text.
+ */
+package eu.exeris.kernel.community.http;
+
+import eu.exeris.kernel.spi.http.HttpHeader;
+import eu.exeris.kernel.spi.http.HttpMethod;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Package-private HPACK-decode accumulator for one HTTP/2 request used by
+ * {@link Http2SessionContext#decodePendingRequest()}.
+ *
+ * <p>Extracted from {@link CommunityHttp2SessionProcessor} in v0.8 Sprint 3
+ * (QA-018a) as one of four seams of the processor's God-class decomposition.
+ * Enforces RFC 7540 §8.1.2.1 pseudo-header ordering (pseudo-headers MUST
+ * precede regular headers), §8.1.2.2 connection-specific header rejection,
+ * and the {@code :method} / {@code :path} pseudo-header requirements.
+ *
+ * <p>Instances are short-lived (one per HEADERS+CONTINUATION block).
+ */
+final class PendingRequestHeaders {
+
+    private String methodToken;
+    private String path;
+    private boolean valid = true;
+    private boolean sawRegularHeader;
+    private final List<HttpHeader> requestHeaders = new ArrayList<>();
+
+    /* default */ void accept(String name, String value) {
+        if (!valid) {
+            return;
+        }
+        if (name.startsWith(":")) {
+            acceptPseudoHeader(name, value);
+            return;
+        }
+        sawRegularHeader = true;
+        if (Http2SessionContext.isConnectionSpecificHeader(name)) {
+            valid = false;
+            return;
+        }
+        requestHeaders.add(new HttpHeader(name, value));
+    }
+
+    /* default */ void invalidate() {
+        valid = false;
+    }
+
+    /* default */ Http2DecodedRequest toDecodedRequest(int streamId) {
+        HttpMethod method = parseHttp2Method(methodToken);
+        String resolvedPath = path == null ? "" : path;
+        boolean requestValid = valid && method != null && !resolvedPath.isEmpty();
+        return new Http2DecodedRequest(streamId, method, resolvedPath, List.copyOf(requestHeaders), requestValid);
+    }
+
+    private void acceptPseudoHeader(String name, String value) {
+        if (sawRegularHeader) {
+            valid = false;
+            return;
+        }
+        switch (name) {
+            case ":method" -> methodToken = value;
+            case ":path" -> path = value;
+            case ":authority", ":scheme" -> {
+                // Accepted but not required for this processing path.
+            }
+            default -> valid = false;
+        }
+    }
+
+    private static HttpMethod parseHttp2Method(String methodToken) {
+        if (methodToken == null || methodToken.isBlank()) {
+            return null;
+        }
+        try {
+            return HttpMethod.valueOf(methodToken);
+        } catch (IllegalArgumentException _) {
+            return null;
+        }
+    }
+}
